Remove dependency to org.eclipse.ui to enable e4 applications
By referencing org.eclipse.ui, pure e4 applications cannot use SWTChart
library as it pulls in all workbench dependencies.

The previous version 0.10.0 published in the org.swtchart namespace did
not have that dependency. In the current code, it is only needed for
the Activator to extend AbstractUIPlugin. As the Activator does not seem
to be used, this change removes it.

Signed-off-by: Andreas Buchen <andreas.buchen@gmail.com>