Fixed tick label not shown if formatted in non-US locale
While updating the tick visibility, AxisTickLabels converts the label
back to a double value. Previously it used Double#parseDouble which is
only parsing values in the Java standard format (local independent).
With this change AxisTickLabels uses the same Format objects that was
used to create the label in the first place.