Remove the dependency on Eclipse ui and runtime plugins
The core swtchart plugin does not use anything beyond pure swt.
The dependency on higher-layer Eclipse bundles like org.eclipse.ui
and org.eclipse.core.runtime is introduced only by the Activator
class which is not really useful in this plugin. Additionally, the
correct base class for a bundle activator is o.s.f.BundleActivator,
not o.e.u.p.AbstractUIPlugin, which makes it additionally
superfluous.
The problem with this dependecy becomes obvious when swtchart core
plugin is used in projects which do not use Eclipse - it enforces
resolving a significant number of otherwise unused bundles.
In order to avoid that problem, the Activator class has been
removed and the manifest adjusted to require only swt bundle.

Signed-off-by: Jaroslaw Kubik <jarek@froglogic.com>