Added an application defined data object associated with each slice
Each slice in a circular chart is represented as a model.Node. When
interacting with the circular chart, the application often need to
retrieve the underlying business object. This change allows to attach
the business object as data object to the node - similar to SWT Widgets
have a data object.